+++ lib/src/main/java/de/fullben/processing/gaia/particles/ParticleWorld.java
@@ -63,7 +63,7 @@ public class ParticleWorld {
    *
    * <p>Note that this method has no effect if the provided duration exceeds the maximum frame
    * duration value specified in the global configuration (available at {@link
-   * Configuration#getConfig()}), as processing the physics for longer durations generally only
+   * Configuration#current()}), as processing the physics for longer durations generally only
    * produces undesirable results.
    *
    * @see #update(double)
@@ -80,7 +80,7 @@ public class ParticleWorld {
    *
    * <p>Note that this method has no effect if the provided duration exceeds the maximum frame
    * duration value specified in the global configuration (available at {@link
-   * Configuration#getConfig()}), as processing the physics for longer durations generally only
+   * Configuration#current()}), as processing the physics for longer durations generally only
    * produces undesirable results.
    *
    * @param durationMillis duration in milliseconds which has elapsed since the last time the world
@@ -105,7 +105,7 @@ public class ParticleWorld {
    */
   private void runPhysics(double duration) {
     // Is the duration ok?
-    if (duration <= 0 || duration > Configuration.getConfig().getMaxFrameDuration()) {
+    if (duration <= 0 || duration > Configuration.current().getMaxFrameDuration()) {
       System.err.println("Attempted to run physics with unsafe frame duration: " + duration + "s");
       return;
     }